refactor: Update badge and update warning asset for orange theme

+++ OUDS/Core/Components/Sources/Indicators/Badge/Internal/BadgeLayout.swift
@@ -26,7 +26,7 @@ struct BadgeLayout {
         case empty(size: OUDSBadge.StandardSize)
 
         /// To display a count on badge
-        case count(value: UInt, size: OUDSBadge.IllustrationSize)
+        case count(value: UInt8, size: OUDSBadge.IllustrationSize)
 
         /// To display an icon. For `OUDSTag.Status.Category.neutral` and `OUDSTag.Status.Category.accent`
         /// the decorative icon must be provided. For other categories, a default icon is already provided.
